requirements.txt
fastapi
uvicorn
rembg
pillow
requests
yt-dlp
python-multipart
jinja2
python-dateutil